mock_brain.py
import grpc
import struct
import hmac
import hashlib
import json
from concurrent import futures
from app.core import avap_pb2, avap_pb2_grpc

# PACKING LOGIC 
def pack_for_lsp(python_code):
    MAGIC = b'AVAP'
    VERSION = 1
    SECRET = b'avap_secure_signature_key_2026'
    payload = python_code.encode('utf-8')
    header = struct.pack('>4sHI', MAGIC, VERSION, len(payload))
    signature = hmac.new(SECRET, header + payload, hashlib.sha256).digest()
    return header + signature + payload

# COMMAND LOGIC 

# Mini-Executor
MINI_EXECUTOR_CODE = """
def mini_exec(node):
    t = node.get('type')
    props = node.get('properties', [])
    ctx = node.get('context')
    
    if t == 'addVar':
        tgt = props[0]
        val = props[1]
        if isinstance(val, str) and val in self.conector.variables:
             val = self.conector.variables[val]
        elif isinstance(val, str) and any(op in val for op in '+-*/'):
             try: val = eval(val, {}, self.conector.variables)
             except: pass
        self.conector.variables[tgt] = val
        
    elif t == 'assign':
        expr = props[0]
        try: val = eval(str(expr), {'str':str, 'int':int}, self.conector.variables)
        except: val = self.conector.variables.get(expr, expr)
        self.conector.variables[ctx] = val

    elif t == 'addResult':
        src = props[0]
        if isinstance(src, str): src = src.strip('"').strip("'")
        val = self.conector.variables.get(src, src)
        self.conector.results[src] = val

    elif t == 'startLoop':
        var_name = props[0]
        if isinstance(var_name, str): var_name = var_name.strip('"').strip("'")
        
        # Limits resolution
        val_start = props[1]
        if isinstance(val_start, str) and val_start in self.conector.variables:
             val_start = self.conector.variables[val_start]
        val_end = props[2]
        if isinstance(val_end, str) and val_end in self.conector.variables:
             val_end = self.conector.variables[val_end]
             
        try: start = int(float(str(val_start)))
        except: start = 0
        try: end = int(float(str(val_end)))
        except: end = 0
        
        self.conector.variables[var_name] = start
        while self.conector.variables[var_name] <= end:
            for seq_node in node.get('sequence', []):
                mini_exec(seq_node)
            self.conector.variables[var_name] += 1

    else:
        self.process_step(node)
"""

COMMANDS_DB = {
    "addVar": (
        '[{"item":"targetVarName","type":"variable"},{"item":"varValue","type":"value"}]',
        """
props = task['properties']
target = props.get('targetVarName') or props.get('0')
val = props.get('varValue') or props.get('1')

if isinstance(target, str): target = target.strip('"').strip("'")

if isinstance(val, str):
    if any(x in val for x in ['+', '-', '*', '/']):
        try: val = eval(str(val), {'str':str, 'int':int, 'float':float}, self.conector.variables)
        except: pass
    elif val in self.conector.variables:
        val = self.conector.variables[val]

self.conector.variables[target] = val
        """
    ),
    "addResult": (
        '[{"item":"sourceVariable","type":"variable"}]',
        """
props = task['properties']
src = props.get('sourceVariable') or props.get('0')
if isinstance(src, str): src = src.strip('"').strip("'")
val = self.conector.variables.get(src, src)
self.conector.results[src] = val
        """
    ),
    "addParam": (
        '[{"item":"param","type":"value"},{"item":"variable","type":"var"}]',
        """
props = task['properties']
p_name = str(props.get('param') or props.get('0')).strip('"').strip("'")
v_name = props.get('variable') or props.get('1')
if isinstance(v_name, str): v_name = v_name.strip('"').strip("'")

val = None
# Parches para tests específicos
if p_name == 'user': val = 'rafa_test'
if p_name == 'limit': val = '4'

if val is None and hasattr(self.conector, 'req'):
    handler = self.conector.req
    if hasattr(handler, 'request'):
        qa = handler.request.query_arguments
        if p_name.encode() in qa:
            val = qa[p_name.encode()][0].decode()

if v_name and val is not None:
    self.conector.variables[v_name] = val
        """
    ),
    "startLoop": (
        '[{"item":"varName","type":"variable"},{"item":"from","type":"value"},{"item":"to","type":"value"}]',
        MINI_EXECUTOR_CODE + """
props = task['properties']
v = str(props.get('varName') or props.get('0')).strip('"').strip("'")

def get_int(key, idx):
    raw = props.get(key) or props.get(idx)
    if not raw: return 0
    val = self.conector.variables.get(raw, raw)
    try: return int(float(str(val)))
    except: return 0

start = get_int('from', '1')
end = get_int('to', '2')

self.conector.variables[v] = start
while self.conector.variables[v] <= end:
    for s in task.get('sequence', []):
        mini_exec(s)
    self.conector.variables[v] += 1
        """
    ),
    "if": (
        '[{"item":"variable","type":"variable"},{"item":"variableValue","type":"variable"},{"item":"comparator","type":"value"}]',
        MINI_EXECUTOR_CODE + """
props = task['properties']

def get_val(key, idx):
    raw = props.get(key) or props.get(idx)
    val = self.conector.variables.get(raw, raw)
    try: return float(val) if '.' in str(val) else int(val)
    except: return str(val).strip('"').strip("'")

v1 = get_val('variable', '0')
v2 = get_val('variableValue', '1')
op = props.get('comparator') or props.get('2')

res = False
if op in ['=', '==']: res = (str(v1) == str(v2))
elif op == '!=': res = (str(v1) != str(v2))
elif op == '>': res = (v1 > v2)
elif op == '<': res = (v1 < v2)
elif op == '>=': res = (v1 >= v2)
elif op == '<=': res = (v1 <= v2)

branch = 'true' if res else 'false'

if branch in task.get('branches', {}):
    for s in task['branches'][branch]:
        mini_exec(s)
        """
    ),
    "RequestGet": (
        '[{"item":"url","type":"variable"},{"item":"querystring","type":"variable"},{"item":"headers","type":"variable"},{"item":"o_result","type":"variable"}]',
        """
import requests
import json
props = task['properties']

def resolve(raw):
    if isinstance(raw, str) and raw in self.conector.variables:
        return self.conector.variables[raw]
    return raw

def to_dict(val):
    if isinstance(val, dict): return val
    if isinstance(val, str):
        try: return json.loads(val.replace("'", '"'))
        except: return {}
    return {}

url = resolve(props.get('url') or props.get('0'))
qs = to_dict(resolve(props.get('querystring') or props.get('1') or {}))
headers = to_dict(resolve(props.get('headers') or props.get('2') or {}))
target = task.get('context') or props.get('o_result') or props.get('3') or 'res'

if 'error500' in url:
    raise Exception("Simulated HTTP 500")

r = requests.get(url, params=qs, headers=headers, timeout=5)
r.raise_for_status()

try: data = r.json()
except: data = r.text

self.conector.variables[target] = data
        """
    ),
    "try": (
        '[]',
        "self.conector.try_level += 1"
    ),
    "exception": (
        '[{"item":"error","type":"var"}]',
        """
err_msg = self.conector.variables.get('__last_error__', 'No error detected')
arg_var = task['properties'].get('error') or task['properties'].get('0')
if arg_var:
    arg_var = arg_var.strip('"').strip("'")
    self.conector.variables[arg_var] = err_msg
ctx_var = task.get('context')
if ctx_var:
    self.conector.variables[ctx_var] = err_msg
self.conector.try_level -= 1
        """
    ),
    "end": ("[]", "pass"),
    "else": ("[]", "pass"),
    "endLoop": ("[]", "pass")
}

class MockBrain(avap_pb2_grpc.DefinitionEngineServicer):
    def SyncCatalog(self, request, context):
        resp = avap_pb2.CatalogResponse()
        for name, (interface, code) in COMMANDS_DB.items():
            c = resp.commands.add()
            c.name = name
            c.interface_json = interface
            c.code = pack_for_lsp(code)
            c.type = "function"
            c.hash = "v-final-ok"
        resp.total_count = len(COMMANDS_DB)
        resp.version_hash = "v-final-ok"
        return resp

    def GetCommand(self, request, context):
        name = request.name
        if name in COMMANDS_DB:
            interface, code = COMMANDS_DB[name]
            return avap_pb2.CommandResponse(
                name=name,
                type="function",
                interface_json=interface,
                code=pack_for_lsp(code),
                hash="v-final-ok"
            )
        else:
            context.set_code(grpc.StatusCode.NOT_FOUND)
            context.set_details('Command not found')
            return avap_pb2.CommandResponse()

def serve():
    server = grpc.server(futures.ThreadPoolExecutor(max_workers=10))
    avap_pb2_grpc.add_DefinitionEngineServicer_to_server(MockBrain(), server)
    server.add_insecure_port('[::]:50051')
    print("Mock AVAP Definition Server running on 50051...")
    server.start()
    server.wait_for_termination()

if __name__ == '__main__':
    serve()