requirements.txt
# Web Server
tornado==6.3.3

# Database & Persistence
asyncpg==0.29.0
psycopg2-binary>=2.9

# Utilities
pyyaml>=6.0
nest-asyncio==1.5.8
requests>=2.31.0

# Security & Data Structure 

python-dotenv>=1.0.0

# gRPC integration
grpcio==1.76.0
grpcio-tools==1.76.0

# Performance
orjson==3.11.5
uvloop==0.22.1